Consolidate three identical types in validation package into a single SchemaOptions type
