+++ internal/validation/schemautil/extract.go
